GraduatedNonConvexity

Documentation for GraduatedNonConvexity.

Introduction

This library is built to solve optimization problems of the form

\[\min_{x} \sum_{i=1}^N \rho_\mu ( r ( y_i, x) )\]

where $\rho_\mu$ is the surrogate of a (non-convex) robust cost function $\rho$.

To do this, it instead solves two problems repeatedly in a loop:

\[x^{k+1} \gets \operatorname{argmin}_x \sum_{i=1}^N w_i^{k} r^2(y_i, x)\]

\[w^{k+1} \gets \operatorname{argmin}_w \sum_{i=1}^N w_i r^2(y_i, x^{k+1}) + \Phi_\rho(w_i)\]

i.e., it solves the non-convex robust optimization as a sequence of weighted least-squares problems. This method is particulary good when step 1 has a closed-form solution or a good solver, and step 2 has analytic solutions.

This library assumes you have a function LSQ_fn that can perform the optimization in step 1. Then, it performs step 2 analytically, for some specific robust cost functions.

The supported robust cost functions are:

Geman McClure:

\[\rho(r) = \frac{\bar{c}^2 r^2}{\bar{c}^2 + r^2}\]

Truncated Least Squares:

\[\rho(r) = \begin{cases} r^2 & \text{if } r^2 \in [0, \bar{c}^2]\\ \bar{c}^2 & \text{if } r^2 \in [\bar{c}^2, \infty) \end{cases}\]

Quick Start

See the example below on how to use the library:

First, we construct the data

using LinearAlgebra

## construct dataset
N=1000

# this is the ground truth parameter we are trying to solve for
β_gt = randn(2)

x = randn(N, 2)
x[:,2] .= 1

# create some noise-free measurements
y = x * β_gt

# add some outliers to 2% of data
for i=1:N
    rand() < 0.02 ? y[i] += 1.0 + rand()  : continue
end

# collect all the necessary data
data = (x, y)

([-0.34968896582759074 1.0; 0.09317097922533679 1.0; … ; -1.2591876089095928 1.0; 0.5888981555315713 1.0], [-0.4060760819947305, -0.4992458799394098, -0.37862260031362654, -0.3924263142251756, -0.388870050300715, -0.3297651449908242, -0.17235393012715958, -0.38230736353669426, -0.35511045388218226, -0.6819379478257174  …  -0.4601485404209124, -0.7210323308906846, -0.27246461925733517, -0.5270759048176012, -0.7330816539684353, 1.2466552831501403, -0.46896475398767673, -0.30284705109578974, -0.2147338625109096, -0.6035379961447224])

Next, we define the weighted least squares solver, and the residual calculation functions

## define weighted least squares solver
function least_sq_solver(w, data)

    X = data[1]
    y = data[2]

    W = diagm(w)

    # analytic expression for weighted least squares
    return (X' * W * X) \ (X' * W * y)
end

## define residual function
function residual_fn(β, data)
  X = data[1]
  y = data[2]
  return y - X * β
end

residual_fn (generic function with 1 method)

And now solve:

using GraduatedNonConvexity

# define maximum residual of inliers
c = 0.3

## Solve using GNC
β_gm = GNC_GM(N, data, least_sq_solver, residual_fn, c; verbose=false) # or
β_tls = GNC_TLS(N, data, least_sq_solver, residual_fn, c; verbose=false)

2-element Vector{Float64}:
 -0.21038208351298127
 -0.47964437520703884

To save allocations, you can also provide in-place versions of the least squares solvers and the residual functions:

function least_sq_solver!(x, w, data)
    ...
end 

function residual_fn!(res, β, data)
    ...
end

and you can call the library in place:

β_gnc = zeros(2) # provide the initial guess for the solution
GNC_GM!(β_gnc, N, data, least_sq_solver!, residual_fn!, c; verbose=false) # or
GNC_TLS!(β_gnc, N, data, least_sq_solver!, residual_fn!, c; verbose=false)

References

See the following paper for details on the method and implementation:

H. Yang, P. Antonante, V. Tzoumas, and L. Carlone,
“Graduated Non-Convexity for Robust Spatial Perception: 
From Non-Minimal Solvers to Global Outlier Rejection”,
IEEE Robotics and Automation Letters (RA-L), 2020.

Documentation

GraduatedNonConvexity.GNC_GM

GraduatedNonConvexity.GNC_GM!

GraduatedNonConvexity.GNC_TLS

GraduatedNonConvexity.GNC_TLS!

GraduatedNonConvexity.GNC_GM! — Method

GNC_GM!(x, w, rs, data, LSQ_fn!, RES_fn!, c)

In-place version of GNC_GM

Inputs:

x: starting point for the robust least squares

w: initial weights vector (data will be overriden)

rs: residuals vector (data will be overriden). Both w and rs need to be of the same length.

data: the data which is to be fit

LSQ_fn!: Assumes LSQ_fn!(x, w, data) updates x in-place with the weighted least squares solution using weights w

RES_fn!: Assumes RES_fn!(rs, x, data) updates rs in-place with the returns residuals given the candidate solution x.

c: the maximum residual of an inlier

Parameters:

(same as GNC_GM)

GraduatedNonConvexity.GNC_GM — Method

GNC_GM(N, data, LSQ_fn, RES_fn, c)

Perform robust least squares given data using the German-McClure cost function.

Inputs:

N: the number of data points (equal to the number of residuals)

data: the data which is to be fit

LSQ_fn: Assumes LSQ_fn(w, data) returns the weighted least squares solution using the weights w

RES_fn: Assumes RES_fn(x, data) returns a vector with the residuals given the candidate solution x.

c: the maximum residual of an inlier

Parameters:

max_iterations=1000

μ_factor=1.4: factor to decrease μ by each iteration.

verbose=true

rtol=1e-6

Terminates when rtol is reached, or μ=1.

GraduatedNonConvexity.GNC_TLS! — Method

GNC_TLS!(x, w, rs, data, LSQ_fn!, RES_fn!, c)

In-place version of GNC_TLS

Inputs:

x: starting point for the robust least squares

w: initial weights vector (data will be overriden)

rs: residuals vector (data will be overriden). Both w and rs need to be of the same length.

data: the data which is to be fit

LSQ_fn!: Assumes LSQ_fn!(x, w, data) updates x in-place with the weighted least squares solution using weights w

RES_fn!: Assumes RES_fn!(rs, x, data) updates rs in-place with the returns residuals given the candidate solution x.

c: the maximum residual of an inlier

Parameters:

(same as GNC_TLS)

GraduatedNonConvexity.GNC_TLS — Method

GNC_TLS(N, data, LSQ_fn, RES_fn, c)

Perform robust least squares given data using the Truncated Least Squares cost function.

Inputs:

N: the number of data points (equal to the number of residuals)

data: the data which is to be fit

LSQ_fn: Assumes LSQ_fn(w, data) returns the weighted least squares solution using the weights w

RES_fn: Assumes RES_fn(x, data) returns a vector with the residuals given the candidate solution x.

c: the maximum residual of an inlier

Parameters:

max_iterations=1000

μ_factor=1.4: factor to increase μ by each iteration.

verbose=true

rtol=1e-6

Terminates when rtol is reached.